# roadrecon/frontend-react/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ROADrecon",
  "version": "2.0.0"
 },
 "paths": {
...
  "/api/policies": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "List Policies",
    "operationId": "list_policies_api_policies_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     },
     {
      "name": "state",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "enabled",
          "reporting",
          "disabled"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "State"
      }
     },
     {
      "name": "block",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Block"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_PolicyRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/policies/affecting/{type}/{id}": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Policies Affecting",
    "description": "Policies that include or exclude the object, directly or through groups and roles.",
    "operationId": "policies_affecting_api_policies_affecting__type___id__get",
    "parameters": [
     {
      "name": "type",
      "in": "path",
      "required": true,
      "schema": {
       "enum": [
        "user",
        "group",
        "role",
        "servicePrincipal",
        "application"
       ],
       "type": "string",
       "title": "Type"
      }
     },
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PolicyMatch"
         },
         "title": "Response Policies Affecting Api Policies Affecting  Type   Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/policies/{id}": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Get Policy",
    "operationId": "get_policy_api_policies__id__get",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PolicyDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/policies/{id}/users": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Policy Users",
    "description": "Users in the policy scope (effect=applies, the default) or excluded from it.",
    "operationId": "policy_users_api_policies__id__users_get",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Id"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     },
     {
      "name": "userType",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "Member",
          "Guest"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Usertype"
      }
     },
     {
      "name": "accountEnabled",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Accountenabled"
      }
     },
     {
      "name": "dirSyncEnabled",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Dirsyncenabled"
      }
     },
     {
      "name": "memberOf",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Users that are members of this group.",
       "title": "Memberof"
      },
      "description": "Users that are members of this group."
     },
     {
      "name": "transitive",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "description": "With memberOf: include members of nested groups.",
       "title": "Transitive"
      },
      "description": "With memberOf: include members of nested groups."
     },
     {
      "name": "ownerOf",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Users that own this object.",
       "title": "Ownerof"
      },
      "description": "Users that own this object."
     },
     {
      "name": "memberOfAu",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Users in this administrative unit.",
       "title": "Memberofau"
      },
      "description": "Users in this administrative unit."
     },
     {
      "name": "mfa",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "none",
          "app",
          "phone",
          "fido",
          "windowsHello"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "none = no strong auth method registered.",
       "title": "Mfa"
      },
      "description": "none = no strong auth method registered."
     },
     {
      "name": "perUserMfa",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "Enabled",
          "Enforced",
          "Disabled"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Perusermfa"
      }
     },
     {
      "name": "excludeMailboxOnly",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "description": "Leave out shared and room mailboxes (the MFA view).",
       "title": "Excludemailboxonly"
      },
      "description": "Leave out shared and room mailboxes (the MFA view)."
     },
     {
      "name": "effect",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "applies",
          "excluded"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Effect"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_UserRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/named-locations": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "List Named Locations",
    "operationId": "list_named_locations_api_named_locations_get",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 500,
       "minimum": 1,
       "default": 50,
       "title": "Page Size"
      }
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Sort"
      }
     },
     {
      "name": "order",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "asc",
        "desc"
       ],
       "type": "string",
       "default": "asc",
       "title": "Order"
      }
     },
     {
      "name": "filter",
      "in": "query",
      "required": false,
      "schema": {
       "type": "array",
       "items": {
        "type": "string"
       },
       "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items.",
       "default": [],
       "title": "Filter"
      },
      "description": "Advanced filters, repeatable: `field:op:value`. For `in`/`notIn` the value is a comma list of URI-encoded items."
     },
     {
      "name": "match",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "all",
        "any"
       ],
       "type": "string",
       "description": "How filters combine.",
       "default": "all",
       "title": "Match"
      },
      "description": "How filters combine."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_NamedLocationRow_"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/named-locations/{id}": {
   "get": {
    "tags": [
     "policies"
    ],
    "summary": "Get Named Location",
    "operationId": "get_named_location_api_named_locations__id__get",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/NamedLocationDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
...
   "Condition": {
    "properties": {
     "key": {
      "type": "string",
      "title": "Key",
      "description": "Key in the policy JSON, e.g. Users, Applications, Locations."
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "include": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Include"
     },
     "exclude": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Exclude"
     }
    },
    "type": "object",
    "required": [
     "key",
     "label",
     "include",
     "exclude"
    ],
    "title": "Condition"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "MatchReason": {
    "properties": {
     "condition": {
      "type": "string",
      "title": "Condition",
      "description": "Condition label, e.g. Users, Directory roles, Resources."
     },
     "via": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Via",
      "description": "Chain from the object to what the policy references, e.g. [group, parent group]. Empty = direct."
     },
     "approximate": {
      "type": "boolean",
      "title": "Approximate"
     },
     "eligibleOnly": {
      "type": "boolean",
      "title": "Eligibleonly",
      "description": "Only through an eligible (PIM) role assignment."
     }
    },
    "type": "object",
    "required": [
     "condition",
     "via",
     "approximate",
     "eligibleOnly"
    ],
    "title": "MatchReason",
    "description": "Why a policy side (include or exclude) matches an object."
   },
...
   "MfaSummary": {
    "properties": {
     "methods": {
      "items": {
       "type": "string",
       "enum": [
        "PhoneAppOTP",
        "PhoneAppNotification",
        "OneWaySms",
        "TwoWayVoiceMobile",
        "TwoWayVoiceOffice",
        "TwoWayVoiceAlternateMobile",
        "Email"
       ]
      },
      "type": "array",
      "title": "Methods"
     },
     "defaultMethod": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "PhoneAppOTP",
         "PhoneAppNotification",
         "OneWaySms",
         "TwoWayVoiceMobile",
         "TwoWayVoiceOffice",
         "TwoWayVoiceAlternateMobile",
         "Email"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Defaultmethod"
     },
     "perUserMfa": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Perusermfa",
      "description": "Legacy per-user MFA state: Enabled, Enforced or null."
     },
     "fido": {
      "type": "integer",
      "title": "Fido"
     },
     "windowsHello": {
      "type": "integer",
      "title": "Windowshello"
     }
    },
    "type": "object",
    "required": [
     "methods",
     "defaultMethod",
     "perUserMfa",
     "fido",
     "windowsHello"
    ],
    "title": "MfaSummary"
   },
   "NamedLocationDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "kind": {
      "type": "string",
      "enum": [
       "ip",
       "country"
      ],
      "title": "Kind"
     },
     "trusted": {
      "type": "boolean",
      "title": "Trusted"
     },
     "ipRanges": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Ipranges"
     },
     "countries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Countries"
     },
     "includeUnknownCountries": {
      "type": "boolean",
      "title": "Includeunknowncountries"
     },
     "policyCount": {
      "type": "integer",
      "title": "Policycount"
     },
     "policies": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Policies",
      "description": "Policies that reference this location."
     },
     "policyMatches": {
      "items": {
       "$ref": "#/components/schemas/PolicyMatch"
      },
      "type": "array",
      "title": "Policymatches"
     },
     "raw": {
      "additionalProperties": true,
      "type": "object",
      "title": "Raw"
     }
    },
    "type": "object",
    "required": [
     "id",
     "displayName",
     "kind",
     "trusted",
     "ipRanges",
     "countries",
     "includeUnknownCountries",
     "policyCount",
     "policies",
     "policyMatches",
     "raw"
    ],
    "title": "NamedLocationDetail"
   },
   "NamedLocationRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "kind": {
      "type": "string",
      "enum": [
       "ip",
       "country"
      ],
      "title": "Kind"
     },
     "trusted": {
      "type": "boolean",
      "title": "Trusted"
     },
     "ipRanges": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Ipranges"
     },
     "countries": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Countries"
     },
     "includeUnknownCountries": {
      "type": "boolean",
      "title": "Includeunknowncountries"
     },
     "policyCount": {
      "type": "integer",
      "title": "Policycount"
     },
     "policies": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Policies",
      "description": "Policies that reference this location."
     }
    },
    "type": "object",
    "required": [
     "id",
     "displayName",
     "kind",
     "trusted",
     "ipRanges",
     "countries",
     "includeUnknownCountries",
     "policyCount",
     "policies"
    ],
    "title": "NamedLocationRow"
   },
...
   "ObjectRef": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "type": {
      "type": "string",
      "enum": [
       "user",
       "group",
       "device",
       "servicePrincipal",
       "application",
       "administrativeUnit",
       "role",
       "policy",
       "namedLocation",
       "keyword",
       "value",
       "unknown"
      ],
      "title": "Type"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "sub": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sub",
      "description": "Secondary identifier: UPN for users, appId for apps and SPs."
     }
    },
    "type": "object",
    "required": [
     "id",
     "type",
     "displayName"
    ],
    "title": "ObjectRef",
    "description": "Every reference to an object, anywhere. `keyword`/`value`/`unknown` are not links."
   },
...
   "Page_NamedLocationRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/NamedLocationRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[NamedLocationRow]"
   },
...
   "Page_PolicyRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PolicyRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[PolicyRow]"
   },
...
   "Page_UserRow_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/UserRow"
      },
      "type": "array",
      "title": "Items"
     },
     "total": {
      "type": "integer",
      "title": "Total"
     },
     "page": {
      "type": "integer",
      "title": "Page"
     },
     "page_size": {
      "type": "integer",
      "title": "Page Size"
     }
    },
    "type": "object",
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[UserRow]"
   },
...
   "PolicyCounts": {
    "properties": {
     "inScope": {
      "type": "integer",
      "title": "Inscope"
     },
     "excluded": {
      "type": "integer",
      "title": "Excluded"
     }
    },
    "type": "object",
    "required": [
     "inScope",
     "excluded"
    ],
    "title": "PolicyCounts"
   },
   "PolicyDetail": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "state": {
      "type": "string",
      "enum": [
       "enabled",
       "reporting",
       "disabled"
      ],
      "title": "State"
     },
     "targetsAllUsers": {
      "type": "boolean",
      "title": "Targetsallusers"
     },
     "targetsAllApps": {
      "type": "boolean",
      "title": "Targetsallapps"
     },
     "block": {
      "type": "boolean",
      "title": "Block"
     },
     "grant": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Grant"
     },
     "grantOperator": {
      "type": "string",
      "enum": [
       "AND",
       "OR"
      ],
      "title": "Grantoperator"
     },
     "sessionControls": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Sessioncontrols"
     },
     "modifiedDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Modifieddatetime"
     },
     "parseError": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parseerror"
     },
     "who": {
      "items": {
       "$ref": "#/components/schemas/Condition"
      },
      "type": "array",
      "title": "Who",
      "description": "Users, workload identities."
     },
     "targets": {
      "items": {
       "$ref": "#/components/schemas/Condition"
      },
      "type": "array",
      "title": "Targets",
      "description": "Applications, user actions, authentication contexts."
     },
     "conditions": {
      "items": {
       "$ref": "#/components/schemas/Condition"
      },
      "type": "array",
      "title": "Conditions",
      "description": "Platforms, locations, client apps, device filters, risks."
     },
     "grantControls": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Grantcontrols",
      "description": "Grant controls and authentication strengths, as value refs."
     },
     "session": {
      "items": {
       "$ref": "#/components/schemas/ObjectRef"
      },
      "type": "array",
      "title": "Session"
     },
     "counts": {
      "$ref": "#/components/schemas/PolicyCounts"
     },
     "raw": {
      "additionalProperties": true,
      "type": "object",
      "title": "Raw"
     }
    },
    "type": "object",
    "required": [
     "id",
     "displayName",
     "state",
     "targetsAllUsers",
     "targetsAllApps",
     "block",
     "grant",
     "grantOperator",
     "sessionControls",
     "modifiedDateTime",
     "parseError",
     "who",
     "targets",
     "conditions",
     "grantControls",
     "session",
     "counts",
     "raw"
    ],
    "title": "PolicyDetail"
   },
   "PolicyMatch": {
    "properties": {
     "policy": {
      "$ref": "#/components/schemas/PolicyRow"
     },
     "effect": {
      "type": "string",
      "enum": [
       "included",
       "excluded"
      ],
      "title": "Effect"
     },
     "included": {
      "items": {
       "$ref": "#/components/schemas/MatchReason"
      },
      "type": "array",
      "title": "Included"
     },
     "excluded": {
      "items": {
       "$ref": "#/components/schemas/MatchReason"
      },
      "type": "array",
      "title": "Excluded"
     }
    },
    "type": "object",
    "required": [
     "policy",
     "effect",
     "included",
     "excluded"
    ],
    "title": "PolicyMatch",
    "description": "One policy that concerns an object. Exclusion wins over inclusion."
   },
   "PolicyRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "state": {
      "type": "string",
      "enum": [
       "enabled",
       "reporting",
       "disabled"
      ],
      "title": "State"
     },
     "targetsAllUsers": {
      "type": "boolean",
      "title": "Targetsallusers"
     },
     "targetsAllApps": {
      "type": "boolean",
      "title": "Targetsallapps"
     },
     "block": {
      "type": "boolean",
      "title": "Block"
     },
     "grant": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Grant"
     },
     "grantOperator": {
      "type": "string",
      "enum": [
       "AND",
       "OR"
      ],
      "title": "Grantoperator"
     },
     "sessionControls": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Sessioncontrols"
     },
     "modifiedDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Modifieddatetime"
     },
     "parseError": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parseerror"
     }
    },
    "type": "object",
    "required": [
     "id",
     "displayName",
     "state",
     "targetsAllUsers",
     "targetsAllApps",
     "block",
     "grant",
     "grantOperator",
     "sessionControls",
     "modifiedDateTime",
     "parseError"
    ],
    "title": "PolicyRow"
   },
...
   "UserRow": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "displayName": {
      "type": "string",
      "title": "Displayname"
     },
     "userPrincipalName": {
      "type": "string",
      "title": "Userprincipalname"
     },
     "accountEnabled": {
      "type": "boolean",
      "title": "Accountenabled"
     },
     "mail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mail"
     },
     "department": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department"
     },
     "jobTitle": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Jobtitle"
     },
     "mobile": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mobile"
     },
     "lastPasswordChangeDateTime": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lastpasswordchangedatetime"
     },
     "dirSyncEnabled": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dirsyncenabled"
     },
     "userType": {
      "type": "string",
      "enum": [
       "Member",
       "Guest"
      ],
      "title": "Usertype"
     },
     "mfa": {
      "$ref": "#/components/schemas/MfaSummary"
     }
    },
    "type": "object",
    "required": [
     "id",
     "displayName",
     "userPrincipalName",
     "accountEnabled",
     "mail",
     "department",
     "jobTitle",
     "mobile",
     "lastPasswordChangeDateTime",
     "dirSyncEnabled",
     "userType",
     "mfa"
    ],
    "title": "UserRow"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}